backend/requirements.txt
fastapi 
uvicorn[standard] #serveur ASGI
pydantic
python-dotenv #pour charger var d'environnement(.env) après
python-jose[cryptography] 
passlib[bcrypt]